refactor: :recycle: avatax-calculate-taxes-map -> adapter
