+++ apps/taxes/src/modules/taxjar/calculate-taxes/taxjar-calculate-taxes-adapter.test.ts
@@ -1,5 +1,3 @@
 import { describe } from "vitest";
 
-describe.todo("TaxJarResponseTransformer", () => {});
-
 describe.todo("TaxJarCalculateTaxesAdapter", () => {});
